backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pydantic-settings
bcrypt
numpy
pandas
scikit-learn
scipy
joblib
email-validator
